v2/autogen_debate_poc/requirements.txt
pyautogen==0.2.16
streamlit==1.29.0
google-generativeai==0.3.2
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
plotly==5.17.0
yfinance==0.2.24
asyncio-throttle==1.0.2
aiohttp==3.9.1